Implement same api with using ES6
