+++ src/index.js
@@ -1,0 +1,4 @@
+import NuclearModule from './NuclearModule'
+
+// to avoid `exports.default` non-sense.
+module.exports = NuclearModule
